+++ index.ts
@@ -1,3 +1,4 @@
 export * from './lib/autocomplete';
+export * from './lib/nearby-search';
 export * from './lib/retrieve';
 export * from './lib/definition';