+++ lib/retrieve.ts
@@ -37,6 +37,7 @@ export async function retrieve(query: IPlaceQuery): Promise<IPlace> {
   const administrativeAreaLevel1 = findComponent('administrative_area_level_1', response.result.address_components);
   const administrativeAreaLevel2 = findComponent('administrative_area_level_2', response.result.address_components);
   const country = findComponent('country', response.result.address_components);
+  const postalCode = findComponent('postal_code', response.result.address_components);
 
   return {
     id: response.result.place_id,
@@ -48,5 +49,6 @@ export async function retrieve(query: IPlaceQuery): Promise<IPlace> {
     administrativeAreaLevel2: administrativeAreaLevel2 ? administrativeAreaLevel2.long_name || administrativeAreaLevel2.short_name : '',
     countryCode: country ? country.short_name : '',
     country: country ? country.long_name : '',
+    postalCode: postalCode ? postalCode.long_name : '',
   };
 }